api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
pydantic==2.5.2
boto3==1.33.6
python-dotenv==1.0.0
pyjwt==2.8.0
passlib==1.7.4
redis==5.0.3